Support TELEC certified deivces.

+++ builder.sh
@@ -10,11 +10,28 @@
 # Copy Kernel and Rootfs to Archive
 # Copy Kernel and Rootfs to TFTP server
 
+while getopts lu OPT; do
+	case $OPT in
+		u)
+			UPDATE="yes"
+			;;
+		l)
+			LINK_LATEST="yes"
+			;;
+		*)
+			exit 0
+			;;
+	esac
+done
+shift $((OPTIND - 1))
+
 DEVICE="$1"
 BUILDER_DIR=$(pwd)
 FIRMWARE_DIR="${BUILDER_DIR}/openipc"
 TIMESTAMP=$(date +"%Y%m%d%H%M")
 VERSION=$(stat -c"%Y" $0)
+FIRMWARE_REPOSITORY="https://github.com/OpenIPC/firmware.git"
+BUILDER_REPOSITORY="https://github.com/hsuenaga/openipc_builder.git"
 
 echo_c() {
     # 30 grey, 31 red, 32 green, 33 yellow, 34 blue, 35 magenta, 36 cyan, 37 white
@@ -67,6 +84,13 @@ copy_to_archive() {
     if $(echo ${DEVICE} | grep -q hi3518ev200_lite); then
         autoup_rootfs
     fi
+
+    if [ "X${LINK_LATEST}" = "Xyes" ]; then
+      LATEST="${BUILDER_DIR}/archive/latest"
+
+      rm -f "${LATEST}"
+      ln -s "${BUILDER_DIR}/archive/${DEVICE}/${TIMESTAMP}" "${LATEST}"
+    fi
 }
 
 copy_to_tftp() {
@@ -105,13 +129,18 @@ tree -C "${ITEM}"
 
 sleep 3
 
-echo_c 33 "\nUpdating Builder"
+echo_c 33 "\nUpdating Builder (${BUILDER_REPOSITORY})"
 git pull
 
-rm -rf openipc
+if [ "X${UPDATE}" != "Xyes" ]; then
+    rm -rf openipc
+else
+    echo_c 33 "Skip cleanup due to -u option."
+fi
+
 if [ ! -d "$FIRMWARE_DIR" ]; then
     echo_c 33 "\nDownloading Firmware"
-    git clone --depth=1 https://github.com/OpenIPC/firmware.git "$FIRMWARE_DIR"
+    git clone --depth=1 "$FIRMWARE_REPOSITORY" "$FIRMWARE_DIR"
     cd "$FIRMWARE_DIR"
 else
     echo_c 33 "\nUpdating Firmware"
